# Change default staking key from RSA 4096 to secp256r1