--- a/network/tracked_ip_test.go
+++ b/network/tracked_ip_test.go
@@ -4,11 +4,62 @@
 package network
 
 import (
+	"net"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
+
+	"github.com/ava-labs/avalanchego/staking"
+	"github.com/ava-labs/avalanchego/utils/ips"
 )
+
+var (
+	ip      *ips.ClaimedIPPort
+	otherIP *ips.ClaimedIPPort
+)
+
+func init() {
+	{
+		cert, err := staking.NewTLSCert()
+		if err != nil {
+			panic(err)
+		}
+		stakingCert, err := staking.ParseCertificate(cert.Leaf.Raw)
+		if err != nil {
+			panic(err)
+		}
+		ip = ips.NewClaimedIPPort(
+			stakingCert,
+			ips.IPPort{
+				IP:   net.IPv4(127, 0, 0, 1),
+				Port: 9651,
+			},
+			1,   // timestamp
+			nil, // signature
+		)
+	}
+
+	{
+		cert, err := staking.NewTLSCert()
+		if err != nil {
+			panic(err)
+		}
+		stakingCert, err := staking.ParseCertificate(cert.Leaf.Raw)
+		if err != nil {
+			panic(err)
+		}
+		otherIP = ips.NewClaimedIPPort(
+			stakingCert,
+			ips.IPPort{
+				IP:   net.IPv4(127, 0, 0, 1),
+				Port: 9651,
+			},
+			1,   // timestamp
+			nil, // signature
+		)
+	}
+}
 
 func TestTrackedIP(t *testing.T) {
 	require := require.New(t)