… network/certs_test.go
… network/test_cert_1.crt
… network/test_cert_2.crt
… network/test_cert_3.crt
… network/test_key_1.key
… network/test_key_2.key
… network/test_key_3.key