@StephenButtolph 2024-05-17
Before:
```
ok  	github.com/ava-labs/avalanchego/network/peer	55.400s	coverage: 56.0% of statements
```

After:
```
ok  	github.com/ava-labs/avalanchego/network/peer	1.546s	coverage: 56.5% of statements
```